# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AcceptInvitationRequest:
    description: Accept invitation request with invite code
    properties:
      invite_code:
        type: string
    required:
    - invite_code
    type: object
  models.AcceptInvitationResponse:
    description: Accept invitation response with membership details
    properties:
      membership_id:
        type: string
      message:
        type: string
      org_id:
        type: string
      role:
        type: string
    type: object
  models.CompleteMultipartUploadRequest:
    description: Multipart upload completion request
    properties:
      parts:
        items:
          $ref: '#/definitions/models.CompletedPart'
        minItems: 1
        type: array
      video_id:
        type: string
    required:
    - parts
    - video_id
    type: object
  models.CompleteMultipartUploadResponse:
    description: Multipart upload completion response
    properties:
      message:
        type: string
      video_url:
        type: string
    type: object
...
  models.DownloadURLResult:
    description: Private download URL result
    properties:
      download_url:
        type: string
      expires_at:
        type: integer
    type: object
  models.ErrorResponse:
    description: Error response with details
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
...
  models.InitiateMultipartUploadRequest:
    description: Multipart upload initiation request
    properties:
      duration_seconds:
        minimum: 1
        type: integer
      file_name:
        maxLength: 255
        type: string
      file_size_bytes:
        minimum: 1
        type: integer
    required:
    - duration_seconds
    - file_name
    - file_size_bytes
    type: object
...
  models.InviteUserRequest:
    description: Invite user request with email and role
    properties:
      email:
        type: string
      org_id:
        type: string
      role:
        enum:
        - user
        - manager
        - admin
        type: string
    required:
    - email
    - org_id
    - role
    type: object
  models.InviteUserResponse:
    description: Invite user response with invitation details
    properties:
      email:
        type: string
      expires_at:
        type: integer
      invitation_id:
        type: string
      invite_code:
        type: string
      role:
        type: string
    type: object
  models.ListInvitationsResponse:
    description: List of invitations for organization
    properties:
      invitations:
        items:
          $ref: '#/definitions/models.InvitationInfo'
        type: array
      total:
        type: integer
    type: object
  models.ListMembersResponse:
    description: List of organization members
    properties:
      members:
        items:
          $ref: '#/definitions/models.MemberInfo'
        type: array
      total:
        type: integer
    type: object
  models.LoginRequest:
    description: Login request with email and password
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    description: Login response with tokens and user info
    properties:
      access_token:
        type: string
      expires_at:
        type: integer
      organizations:
        items:
          $ref: '#/definitions/models.OrganizationInfo'
        type: array
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/models.UserInfo'
    type: object
  models.LogoutRequest:
    description: Logout request with refresh token
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.LogoutResponse:
    description: Logout response with message
    properties:
      message:
        type: string
    type: object
...
  models.OrganizationInfo:
    description: Organization information for user
    properties:
      name:
        type: string
      org_id:
        type: string
      role:
        type: string
    type: object
  models.PublishVideoRequest:
    description: Publish video request
    properties:
      video_id:
        type: string
    required:
    - video_id
    type: object
  models.PublishVideoResult:
    description: Publish video result
    properties:
      message:
        type: string
      public_url:
        type: string
    type: object
  models.RefreshTokenRequest:
    description: Refresh token request
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.RefreshTokenResponse:
    description: Refresh token response with new tokens
    properties:
      access_token:
        type: string
      expires_at:
        type: integer
      refresh_token:
        type: string
    type: object
  models.RegisterRequest:
    description: Registration request with user details
    properties:
      email:
        type: string
      full_name:
        type: string
      organization_name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  models.RegisterResponse:
    description: Registration response with user ID and message
    properties:
      message:
        type: string
      user_id:
        type: string
    type: object
  models.SearchVideosResponse:
    description: Search videos response
    properties:
      total_count:
        type: integer
      videos:
        items:
          $ref: '#/definitions/models.Video'
        type: array
    type: object
  models.UpdateMemberRoleRequest:
    description: Update member role request
    properties:
      new_role:
        enum:
        - user
        - manager
        - admin
        type: string
    required:
    - new_role
    type: object
  models.UpdateProfileRequest:
    description: Profile update request with full name
    properties:
      full_name:
        type: string
    required:
    - full_name
    type: object
  models.UserInfo:
    description: User profile information
    properties:
      created_at:
        type: integer
      email:
        type: string
      email_verified:
        type: boolean
      full_name:
        type: string
      org_id:
        type: string
      role:
        type: string
      updated_at:
        type: integer
      user_id:
        type: string
    type: object
  models.VerifyEmailRequest:
    description: Email verification request with email and code
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
  models.VerifyEmailResponse:
    description: Email verification response with success status
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  models.Video:
    description: Video information
    properties:
      duration_seconds:
        type: integer
      file_name:
        type: string
      file_size_bytes:
        type: integer
      upload_status:
        type: string
      uploaded_at:
        type: integer
      video_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for SellerProof video management platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SellerProof API
  version: "1.0"
paths:
  /api/v1/organization/invitations:
    get:
      description: List all invitations for organization (admin/manager only)
      parameters:
      - description: Organization ID
        in: query
        name: org_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListInvitationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List invitations
      tags:
      - organization
...
  /api/v1/organization/invitations/accept:
    post:
      consumes:
      - application/json
      description: Accept invitation to organization using invite code
      parameters:
      - description: Accept invitation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AcceptInvitationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AcceptInvitationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Accept invitation
      tags:
      - organization
  /api/v1/organization/invite:
    post:
      consumes:
      - application/json
      description: Invite user to organization with specific role (admin/manager only)
      parameters:
      - description: Invite user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.InviteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InviteUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Invite user to organization
      tags:
      - organization
  /api/v1/organization/members:
    get:
      description: List all members of organization (admin only)
      parameters:
      - description: Organization ID
        in: query
        name: org_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListMembersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List organization members
      tags:
      - organization
  /api/v1/organization/members/{user_id}:
    delete:
      description: Remove member from organization (admin only)
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove member
      tags:
      - organization
  /api/v1/organization/members/{user_id}/role:
    put:
      consumes:
      - application/json
      description: Update organization member role (admin only)
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Update member role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateMemberRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update member role
      tags:
      - organization
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: User login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user and invalidate refresh token
      parameters:
      - description: Logout request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LogoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - auth
  /auth/profile:
    get:
      consumes:
      - application/json
      description: Get current user profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserInfo'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - auth
    put:
      consumes:
      - application/json
      description: Update current user profile information
      parameters:
      - description: Profile update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token using refresh token
      parameters:
      - description: Refresh token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email, password, full name and optional
        organization name
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verify user email with verification code sent to email
      parameters:
      - description: Email verification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VerifyEmailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Verify user email
      tags:
      - auth
...
  /video:
    get:
      consumes:
      - application/json
      description: Get video information by video ID
      parameters:
      - description: Video ID
        in: query
        name: video_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Video'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get video information
      tags:
      - video
  /video/download:
    get:
      consumes:
      - application/json
      description: Get temporary presigned URL for private video download (1 hour)
      parameters:
      - description: Video ID
        in: query
        name: video_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DownloadURLResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download private video
      tags:
      - video
  /video/publish:
    post:
      consumes:
      - application/json
      description: Publish video to public bucket (admin/manager only)
      parameters:
      - description: Publish video request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PublishVideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PublishVideoResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Publish video
      tags:
      - video
  /video/search:
    get:
      consumes:
      - application/json
      description: Search videos with query and pagination
      parameters:
      - description: Search query
        in: query
        name: query
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SearchVideosResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search videos
      tags:
      - video
  /video/upload/complete:
    post:
      consumes:
      - application/json
      description: Complete multipart upload and create video record
      parameters:
      - description: Multipart upload completion request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CompleteMultipartUploadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CompleteMultipartUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Complete multipart upload
      tags:
      - video
  /video/upload/initiate:
    post:
      consumes:
      - application/json
      description: Initiate multipart upload for video file
      parameters:
      - description: Multipart upload initiation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.InitiateMultipartUploadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InitiateMultipartUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Initiate multipart upload
      tags:
      - video
  /video/upload/urls:
    post:
      consumes:
      - application/json
      description: Get presigned URLs for multipart upload parts
      parameters:
      - description: Part upload URLs request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GetPartUploadURLsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetPartUploadURLsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get part upload URLs
      tags:
      - video
securityDefinitions:
  BearerAuth:
    description: Bearer token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"